adding file

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -9,12 +9,12 @@
 
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	size_t i = 0;
+	int i = 0;
 
 	if (array == NULL || action == NULL)
 		exit(1);
 
-	for (i = 0; i < size; i++)
+	for (i = 0; i < (int) size; i++)
 	{
 		action(array[i]);
 	}
